# server/internal/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bastion API
  description: |
    API documentation for Bastion bot and webhook integrations.

    ## Authentication

    **Bot tokens** use the `Bot` scheme:
    ```
    Authorization: Bot bot_xxxxxxxxxxxx
    ```

    Bot tokens are issued when creating a bot in Server Settings > Integrations.
    They are shown once — copy and store securely. You can regenerate a token
    at any time (which invalidates the previous one).

    **Webhook tokens** are embedded in the URL and require no header:
    ```
    POST /api/v1/webhooks/{webhookID}/{token}
    ```

    **User JWTs** use the `Bearer` scheme (for management endpoints):
    ```
    Authorization: Bearer <jwt>
    ```

    ## WebSocket

    Bots can connect to the real-time WebSocket for live events:
    ```
    GET /api/v1/ws?token=bot_xxxxxxxxxxxx
    ```

    ## Rate Limits

    | Scope | Limit |
    |-------|-------|
    | Webhook execution | 30 requests/min per IP |
    | Protected API (global) | 120 requests/min per user |
    | Message send | 10 messages/10s per user |
    | File upload | 5 uploads/min per user |
    | Bulk import | 10 requests/min per user |

    When rate-limited, the server returns `429 Too Many Requests`.

  version: "1.0"
  contact:
    name: Bastion
    url: https://bastions.org
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /servers/{serverID}/bots/{botID}/commands:
    post:
      tags: [Interactions]
      summary: Register command
      description: |
        Register a slash command (or context menu command) for a bot.
        Requires bot token auth (bot must belong to the server).
      operationId: registerCommand
      security:
        - BotAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommandRequest'
            example:
              name: "ping"
              description: "Responds with pong"
              type: 1
      responses:
        '201':
          description: Command created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationCommand'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Command with that name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [Interactions]
      summary: List bot commands
      description: List all commands registered by a specific bot.
      operationId: listBotCommands
      security:
        - BotAuth: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
      responses:
        '200':
          description: Command list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationCommand'
  /servers/{serverID}/bots/{botID}/commands/{commandID}:
    patch:
      tags: [Interactions]
      summary: Update command
      description: Update a command's description or options.
      operationId: updateCommand
      security:
        - BotAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
        - name: commandID
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application command ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommandRequest'
      responses:
        '200':
          description: Command updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationCommand'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Interactions]
      summary: Delete command
      description: Delete a registered command.
      operationId: deleteCommand
      security:
        - BotAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
        - $ref: '#/components/parameters/botID'
        - name: commandID
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Application command ID
      responses:
        '200':
          description: Command deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /servers/{serverID}/commands:
    get:
      tags: [Interactions]
      summary: List server commands
      description: |
        List all slash commands and context menu commands available in a server
        (from all bots). Requires server membership.
      operationId: listServerCommands
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serverID'
      responses:
        '200':
          description: All commands in the server
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationCommand'
...
components:
  securitySchemes:
    BotToken:
      type: apiKey
      in: header
      name: Authorization
      description: 'Bot token: `Bot bot_xxxxxxxxxxxx`'
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login
  parameters:
    serverID:
      name: serverID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Server ID
...
    botID:
      name: botID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Bot ID
...
  schemas:
...
    CommandOption:
      type: object
      properties:
        name:
          type: string
          description: Option name
        description:
          type: string
        type:
          type: integer
          description: "3=STRING, 4=INTEGER, 5=BOOLEAN, 6=USER, 7=CHANNEL"
        required:
          type: boolean
        choices:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: string
    ApplicationCommand:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serverId:
          type: string
          format: uuid
        botId:
          type: string
          format: uuid
        type:
          type: integer
          description: "1=CHAT_INPUT, 2=USER, 3=MESSAGE"
        name:
          type: string
        description:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/CommandOption'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateCommandRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: "Lowercase alphanumeric + hyphens/underscores, 1-32 chars"
          example: "ping"
        description:
          type: string
          maxLength: 100
          example: "Responds with pong"
        type:
          type: integer
          description: "1=CHAT_INPUT (default), 2=USER, 3=MESSAGE"
          default: 1
        options:
          type: array
          items:
            $ref: '#/components/schemas/CommandOption'
    UpdateCommandRequest:
      type: object
      properties:
        description:
          type: string
          maxLength: 100
        options:
          type: array
          items:
            $ref: '#/components/schemas/CommandOption'
...
    DeleteResponse:
      type: object
      properties:
        status:
          type: string
          example: "deleted"
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "name is required and must be 1-100 characters"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "resource not found"